Dev.1.0.0 fix merge (#13)
* fix slow shutdown issue

* fix core dump when shutdown (workaround)

* feat: Implement min difficulty (max ease)

* fix: bug related bnMaximumTarget

* fix: update variable nBigLinearTrailingAverageLengt correctly

* test: allow update nDifficulty at yacoin.conf to support test faster

* fix: update method save latest block reward, no need to save in database

* Merge branch 'fix-shutdown-rpc' into dev.1.0.0

* fix: error when build on Ubuntu 16.04

* Support 64bit nTime for CBlockIndex, CDiskBlockIndex

+ Write/Read 64bit time in leveldb
+ Modify other places using nTime of CBlockIndex

* Support write/read 64-bit nTime of Cblock in blk*.dat

* Modify all places which use Cblock nTime for hash calculation

Modify all places which use Cblock nTime for hash calculation. Otherwise, nTime size change leads to different hash when calculating hash for old block

* Modify FormatHashBuffers which reverse block header in 4 byte chunks

* Modify CTransaction nTime

* fixup! Modify all places which use Cblock nTime for hash calculation

* fixup2! Modify all places which use Cblock nTime for hash calculation

* fixup! Modify CTransaction nTime

* Only extend size of nTime at the time happening hardfork

* Workaround to avoid extend size of nTime for genesis block in testnet

* Fix wrong hash calculation due to data padding

* Change block/tx version at the time happenning hardfork

Block/tx version helps us to separate between 32bit-nTime block/tx and 64bit-nTime block/tx
Only extend size of block/tx nTime at the time happening hardfork

* fix: Bug relate to UTC time fix

+++ src/wallet.cpp
@@ -581,8 +581,8 @@ bool CWallet::AddToWallet(const CWalletTx& wtxIn)
             {
                 if (mapBlockIndex.count(wtxIn.hashBlock))
                 {
-                    unsigned int latestNow = wtx.nTimeReceived;
-                    unsigned int latestEntry = 0;
+                    int64_t latestNow = wtx.nTimeReceived;
+                    int64_t latestEntry = 0;
                     {
                         // Tolerate times up to the last timestamp in the wallet not more than 5 minutes into the future
                         int64_t latestTolerated = latestNow + 300;
@@ -613,7 +613,7 @@ bool CWallet::AddToWallet(const CWalletTx& wtxIn)
                         }
                     }
 
-                    unsigned int& blocktime = mapBlockIndex[wtxIn.hashBlock]->nTime;
+                    int64_t& blocktime = mapBlockIndex[wtxIn.hashBlock]->nTime;
                     wtx.nTimeSmart = std::max(latestEntry, std::min(blocktime, latestNow));
                 }
                 else
@@ -1670,7 +1670,7 @@ int64_t CWallet::GetWatchOnlyNewMint() const
     return nTotal;
 }
 
-bool CWallet::SelectCoinsMinConf(int64_t nTargetValue, unsigned int nSpendTime, int nConfMine, int nConfTheirs, vector<COutput> vCoins, set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet) const
+bool CWallet::SelectCoinsMinConf(int64_t nTargetValue, int64_t nSpendTime, int nConfMine, int nConfTheirs, vector<COutput> vCoins, set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet) const
 {
     setCoinsRet.clear();
     nValueRet = 0;
@@ -1779,7 +1779,7 @@ bool CWallet::SelectCoinsMinConf(int64_t nTargetValue, unsigned int nSpendTime, 
     return true;
 }
 
-bool CWallet::SelectCoins(int64_t nTargetValue, unsigned int nSpendTime, set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet, const CCoinControl* coinControl) const
+bool CWallet::SelectCoins(int64_t nTargetValue, int64_t nSpendTime, set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet, const CCoinControl* coinControl) const
 {
     vector<COutput> vCoins;
     
@@ -1804,7 +1804,7 @@ bool CWallet::SelectCoins(int64_t nTargetValue, unsigned int nSpendTime, set<pai
 }           // OK, quick, without looking at SelectCoinsMinConf(), WTF are 1, 6 - 1, 1 - 0, 1????
 
 // Select some coins without random shuffle or best subset approximation
-bool CWallet::SelectCoinsSimple(int64_t nTargetValue, int64_t nMinValue, int64_t nMaxValue, unsigned int nSpendTime, int nMinConf, set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet) const
+bool CWallet::SelectCoinsSimple(int64_t nTargetValue, int64_t nMinValue, int64_t nMaxValue, int64_t nSpendTime, int nMinConf, set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet) const
 {
     vector<COutput> vCoins;
     AvailableCoinsMinConf(vCoins, nMinConf, nMinValue, nMaxValue);
